SnapdragonCamera: Fix shutter button clicks in rapid succession getting ignored
Due to this check, the shutter button would be disabled until each photo
would finish processing. This results in the shutter button getting
disabled for quite a long time when ZSL isn't used, forcing the user to
wait until the previous shot is finished before requesting a new shot.

Fix the HDR check so that shutter button clicks in quick succession work.
There is code in place to handle button clicks in rapid succession, so this
is fine.

Change-Id: I7b6fc1cb660dba2e505238560f932e98adc1f2ce
Signed-off-by: Alex Naidis <alex.naidis@linux.com>