Snap: Fix shutter button clicks in rapid succession getting ignored
Due to this check, the shutter button would be disabled until each photo
would finish processing. This results in the shutter button getting
disabled for quite a long time when ZSL isn't used, forcing the user to
wait until the previous shot is finished before requesting a new shot.

Fix the HDR check so that shutter button clicks in quick succession work.
There is code in place to handle button clicks in rapid succession, so this
is fine.

Change-Id: Ifc36a22e4ec3a1be24e5a9b40c3fc682e82605b3
Signed-off-by: Alex Naidis <alex.naidis@linux.com>