SnapdragonCamera: Make sure photos get saved when app exits
Although we have a media save service that runs in the background, it's
not useful because the JPEG callback function runs in the UI thread, which
gets killed when the app exits. When this happens, pictures that have not
been processed through the JPEG callback before the app exits are
permanently lost. To reproduce, simply capture a photo and leave the app
before the photo is saved; this will cause the captured photo to never be
saved.

To fix this, place the takePicture() callback routines into a separate
thread and stall the UI thread until all photos are saved when the app
exits. This solution is messy, but it works.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

* Also fix all resulting FCs.

Change-Id: Ie6d1ab26d594c941cd7ea4907fe9e656958f34e6
Signed-off-by: Alex Naidis <alex.naidis@linux.com>